Fix Livewire component error
Create a new post and the post is successfully seeded in the database,
but devtool reports this error:

Uncaught Could not find Livewire component in DOM tree

When we wrap Livewire components inside a Blade component, Livewire may
lose track of wire:id, component boundaries, or DOM node relationships,
because Blade components use {{ $slot }} which can interfere with how
Livewire hydrates and diffs the DOM.

When a new post is added, the post-manager component is updated with the
new post.  But the post-list component inside the Blade component is not
updated with the new post, making Livewire lose tracking of its wire:id,
component boundaries or DOM node relationships.  Hence the error.

Fix this issue by merging the post-manager component into the post-list
component.  Now when we add a new post, it is seeded in the database.
The post-list is happily updated and Livewire has all the info of the
pertinent component even it is inside the Blade component.