Build editPost method to fill the form
Build the method to pre-fill the form with the values of the pertinent
post.  Add editPostId property for the edit mode.  Also build the update
policy to let the owner edit the post.

User clicks "Edit" on a post and editPost($id) runs:

1) Loads the post from DB
2) Fills the form fields: $this->title, $this->content
3) Sets $this->editingPostId = $id

Now, the form is pre-filled and it's in edit mode because editingPostId
property is not null.

Next, the user can proceed to update by overriding the pertinent post.